### 2020-02-22 #2115 @oliviaguest
Also @deniederhut is this a library for dealing with linguistic data or data in general? Sorry I am skimming and it looks like specifically for corpora of text but I am struggling to see it explicitly stated in the docs https://niacin.readthedocs.io/en/latest/? I might be missing something though, either way, please feel free (if you think it's useful) to link me to the right place so I can read more.
