### 2020-02-23 #2115 @deniederhut
😅 the intention was to also include algorithms for image data and timeseries, but I haven't had a chance to write any of that. Right now it is just text augmentation.

As for reviewers, I might suggest the following:
- rlskoeser
- adammichaelwood
- sara-02
- mbod
