### 2020-02-23 #2115 @oliviaguest
> 😅 the intention was to also include algorithms for image data and timeseries, but I haven't had a chance to write any of that. Right now it is just text augmentation.

Thanks for clarifying. Can you make some edits to make this current state and plan clear, please?
