### 2020-02-23 5b5595a
DOC: clarify library use for text data
