### 2020-03-02 #2115 @deniederhut
Hey! I modified the titles of both the paper and the README last week per your request - you can see the [changes here](https://github.com/deniederhut/niacin/commit/5b5595a069d5233985e8e842353fcf3ca4af8561).

The body of the paper already stated this focus explicitly:

> Here, we introduce niacin, a python library for performing a large set of commonly-used text transformations

Were there further changes you would like to see?
