### 2020-03-04 #2136 @whedon
[REVIEW]: niacin: A Python package for data enrichment
**Submitting author:** @deniederhut (<a href="http://orcid.org/0000-0002-9995-4378">Dillon Niederhut</a>)
**Repository:** <a href="https://github.com/deniederhut/niacin" target ="_blank">https://github.com/deniederhut/niacin</a>
**Version:** 0.2.0
