### 2020-03-04 #2136 @oliviaguest
👋 @BayesForDays and @sara-02 — feel free to leave any comments here you have about the manuscript and if you have code-related comments please use the repository's issues (https://github.com/deniederhut/niacin) and link to them from here. The main bulk of your review involves the list above. 😺

👉 If you can both give a rough ETA for when you can complete your reviews that would be very helpful.
