### 2020-03-04 #2136 @sara-02
> wave @BayesForDays and @sara-02 — feel free to leave any comments here you have about the manuscript and if you have code-related comments please use the repository's issues (https://github.com/deniederhut/niacin) and link to them from here. The main bulk of your review involves the list above. smiley_cat
>
> point_right If you can both give a rough ETA for when you can complete your reviews that would be very helpful.
I should have my first set of reviews by March 15th.
