### 2020-03-06 #2136 @sara-02
@deniederhut Will it be possible to include some sort of feature/architecture comparison between `niachin, EDA and noisemix`?
This is w.r.t to the following parameter required regarding the paper
`State of the field: Do the authors describe how this software compares to other commonly-used packages?`
In the current paper, we are only mentioning the following
`here are a small number of libraries that contain text
augmentation functions, for example noisemix and EDA (Wei & Zou, 2019), but these tend
to be associated with conference papers and have little documentation, or only implement a
small subset of common transforms.`
I think we can expand on this WDYT?
