### 2020-03-10 #2136 @deniederhut
> Will it be possible to include some sort of feature/architecture comparison

Sure thing 😄
