### 2020-03-10 a30328b
DOC: add import statement to quickstart examples
Closes https://github.com/deniederhut/niacin/issues/12
