### 2020-03-10 23a1b3b
DOC: adds instructions for testing
Closes https://github.com/deniederhut/niacin/issues/11
