### 2020-03-10 9e26f89
DOC: adds more information about transforms
And a rough equivalency table between libraries.
