### 2020-03-24 #2136 @deniederhut
Hey! Sorry for holding things up -- things have been a bit hectic. I've merged the PRs with the requested changes, and am presently addressing some tickets that @sara-02 created. Thanks so much for your help Sarah!

@BayesForDays this is ready for review whenever you are. No rush 😄
