### 2020-03-25 c782e9e
DOC: adds docstrings for private functions
Closes https://github.com/deniederhut/niacin/issues/18
