### 2020-03-25 27c4979
DOC: fix docstring for add_whitespace
Closes https://github.com/deniederhut/niacin/issues/16
