### 2020-04-14 5b628f6
ENH: auto-download wordnet data
