### 2020-05-09 b3a8189
DOC: adds extras to install docs
