### 2020-05-09 54202cc
CI: quiet pip and wait for downloads
