+++ .travis.yml
@@ -46,10 +46,12 @@ before_install :
     - conda info -a
     - conda create -q -n test-environment python=$PY_MAJOR_MINOR scipy
     - source activate test-environment
+    - if [ "$TRAVIS_OS_NAME" = "osx" ]; then export CC=clang;export CXX=clang++;export CFLAGS='-stdlib=libc++'; fi
 
 install :
-    - pip install pytest pytest-cov codecov
-    - python setup.py install
+    - pip install -qqq pytest pytest-cov codecov
+    - pip install -qqq .[all]
+    - travis_wait python scripts/fairseq_helper.py
 
 script : py.test --cov=niacin
 
