### 2020-05-11 #2136 @BayesForDays
Hi @deniederhut, I'm checking out the paper right now and I had a couple of questions/comments. One concern is about the wordnet functionality -- and in particular your motivations for using `nltk` to access wordnet. Do you think it makes sense to specify a specific version of nltk?

On a conceptual level, I think it would be helpful for readers if you could elaborate a bit on the wordnet functions (e.g. `add_hypernyms`), as well as include a citation about whether these _specific types_ of negative samples are useful for modeling. I personally think the wordform transformation is the most useful one and requires the least explanation, partly because it's also the most analogous to the computer vision work that I'm aware of.

In the paper, it would be helpful to somehow graphically mark the changes that take place when transforming the "original" text into the "augmented" text as well, especially for the hyper/hyponym example.

Finally, I think it makes sense (unless the package is more general) to specify that the language this package assumes is English. If the package can be made to be more general (e.g. different language versions of wordnet) specifying this somehow or allowing that to be passed in as a parameter would be a nice "to-do."
