### 2020-05-11 #2136 @BayesForDays
I have a handful of questions about the API, nothing major I think.

* Some additional documentation explaining the main modules (e.g. `text`) would be very helpful. For example, it wasn't clear to me why `text` has to be imported rather than it being easy to import a function like `add_hypernyms` directly from `niacin`. The `text` import could also be clarified in the documentation and the paper (which seems to assume that the reader has done something like `from niacin.text import *`).
* Perhaps you could add some docstrings that return some information about what functions are contained in the modules as well?
