### 2020-05-14 6b40534
ENH: adds char insert/delete
