### 2020-05-17 9b0372a
TST: adds tests for swaps
