### 2020-05-17 71996ca
ENH: adds word/character flipping
