### 2020-05-19 #2136 @sara-02
Hey @deniederhut I recently came across some more references that you can add to your literature review and introduction if they not already spoken about.
https://github.com/makcedward/nlpaug
https://github.com/QData/TextAttack
Also, an interesting blog giving an overview of the work you aim to achieve:
https://amitness.com/2020/05/data-augmentation-for-nlp/
