### 2020-05-19 #2136 @deniederhut
Thanks for all the feedback @BayesForDays ! I've been working on adding some new transformations to flesh out the capabilities of the library a bit and those PRs should get merged in soon. I'll turn my attention to the issues you raised next (good catch on the nltk version btw -- it's in requirements.txt but not the setup file 😬). To address this one briefly here:

>  I think it makes sense (unless the package is more general) to specify that the language this package assumes is English

Totally. Cross-lingual support is hard, because some of these transformations don't really make sense for other languages, or don't have similar resources available. Maybe the best approach here is to have something like `from niacin.text import en` -- which will have the added benefit of not loading in resources during runtime for a language that's not being used.

And thanks for the references @sara-02 ! That nlpaug library has some interesting stuff in it. I'm not sure I'll be able add similar functionality here in a timely fashion 😅
