### 2020-05-25 ccb39b4
CLN: reoganizes text module
To put English language functions behind text.en module. Also
fixes some rst rendering issues with docstrings.
