+++ doc/source/niacin.text.en.rst
@@ -1,0 +1,40 @@
+niacin.text.en
+==============
+
+The niacin.text.en package contains functions for augmenting English text data.
+
+Submodules
+----------
+
+niacin.text.en.char module
+--------------------------
+
+.. automodule:: niacin.text.en.char
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+niacin.text.en.sentence module
+------------------------------
+
+.. automodule:: niacin.text.en.sentence
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+niacin.text.en.word module
+--------------------------
+
+.. automodule:: niacin.text.en.word
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: niacin.text.en
+    :members:
+    :undoc-members:
+    :show-inheritance:
