### 2020-05-25 4993a7c
DOC: adds text.en user guide
