+++ doc/source/user_guide/index.rst
@@ -1,0 +1,8 @@
+niacin user guide
+=================
+
+.. toctree::
+   :maxdepth: 3
+
+   text_en
+
