### 2020-05-26 7db1a5c
CI: update imports in model downloader
