### 2020-05-26 #2136 @deniederhut
Okay, I think I have addressed the comments here. Let me know what you think!

@sara-02 I noticed that you closed the issue you raised about support for other languages. Do you think it would make sense to leave it open in case someone else has a similar question?
