### 2020-06-06 #2136 @kyleniemeyer
Hi @deniederhut, just a few minor copy edits:

- In the second paragraph, could you define NLP on its first use?
- Also, please add a comma after "e.g."
- In paragraph 4, I think that it should be "disrupts" in the sentence starting with "Adding nuisance characaters..."
- on second page, in the paragraph starting with "When possible, function documentation...", there is a missing "to" between "how apply"
- also in that paragraph, "empirical" is misspelled
- in the references, there are a few that need {} around words to ensure capitalization. The Miller 1995 reference needs the DOI (10.1145/219717.219748), and the Morris 2020 reference is missing some info

Let me know here when you've made those changes and I'll publish the article!
