### 2020-06-09 #2136 @deniederhut
Okay, typos and capitalization fixed. I have a question about:

> the Morris 2020 reference is missing some info

The bibliographic data is what was provided by the authors. The citation export on arxiv doesn't provide anything more, except the arxiv-specific subcategory. Is this what you meant?
