### 2020-06-09 #2136 @kyleniemeyer
@deniederhut for the Morris reference, I think the problem is that none of the arXiv info is showing properly. I'm not totally sure if our paper parser handles this, but could you try adding these fields?

```
archivePrefix = {arXiv},
eprint = {2005.05909},
primaryClass = {cs.CL},
```

If that doesn't work, then we can hack it by adding this field: `howpublished = {\tt arXiv:2005.05909 [cs.CL]}`
