### 2020-06-11 #2136 @deniederhut
Ah maybe a capitalization issue in the field names? It should be fixed as of https://github.com/deniederhut/niacin/pull/29
